fusion: Restructure fusion engine
Split the fusion.py module into several modules, which improves understanding. Improve documentation as well.
