+++ pyop2/base.py
@@ -166,7 +166,7 @@ class ExecutionTrace(object):
         self._trace = new_trace
 
         if configuration['loop_fusion']:
-            from fusion import fuse, lazy_trace_name
+            from fusion.interface import fuse, lazy_trace_name
             to_run = fuse(lazy_trace_name, to_run)
         for comp in to_run:
             comp._run()
