--- pyop2/fusion.py
@@ -1,1918 +1,0 @@
-# This file is part of PyOP2
-#
-# PyOP2 is Copyright (c) 2012, Imperial College London and
-# others. Please see the AUTHORS file in the main source directory for
-# a full list of copyright holders.  All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions
-# are met:
-#
-#     * Redistributions of source code must retain the above copyright
-#       notice, this list of conditions and the following disclaimer.
-#     * Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in the
-#       documentation and/or other materials provided with the distribution.
-#     * The name of Imperial College London or that of other
-#       contributors may not be used to endorse or promote products
-#       derived from this software without specific prior written
-#       permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTERS
-# ''AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT HOLDERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
-# INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
-# HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
-# STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
-# OF THE POSSIBILITY OF SUCH DAMAGE.
-
-"""OP2 backend for fusion and tiling of parloops."""
-
-from contextlib import contextmanager
-from decorator import decorator
-from collections import OrderedDict
-from copy import deepcopy as dcopy, copy as scopy
-from itertools import groupby
-import os
-import sys
-
-from base import *
-import base
-import compilation
-import sequential
-import host
-from backends import _make_object
-from caching import Cached
-from profiling import timed_region
-from logger import warning, info as log_info
-from mpi import MPI, collective
-from utils import flatten, strip, as_tuple
-
-import coffee
-from coffee import base as ast
-from coffee.utils import ast_make_alias, ItSpace
-from coffee.visitors import FindInstances, SymbolReferences
-
-
-try:
-    """Is SLOPE accessible ?"""
-    sys.path.append(os.path.join(os.environ['SLOPE_DIR'], 'python'))
-    import slope_python as slope
-    os.environ['SLOPE_METIS']
-
-    # Set the SLOPE backend
-    backend = os.environ.get('SLOPE_BACKEND')
-    if backend not in ['SEQUENTIAL', 'OMP']:
-        backend = 'SEQUENTIAL'
-    if MPI.parallel:
-        if backend == 'SEQUENTIAL':
-            backend = 'ONLY_MPI'
-        if backend == 'OMP':
-            backend = 'OMP_MPI'
-    slope.set_exec_mode(backend)
-    log_info("SLOPE backend set to %s" % backend)
-except:
-    warning("Couldn't locate SLOPE, no tiling possible. Check SLOPE_{DIR,METIS} env vars")
-    slope = None
-
-
-lazy_trace_name = 'lazy_trace'
-"""The default name for sequences of par loops extracted from the trace produced
-by lazy evaluation."""
-
-
-class FArg(sequential.Arg):
-
-    """An Arg specialized for kernels and loops subjected to any kind of fusion."""
-
-    def __init__(self, arg, gather=None, c_index=False):
-        """Initialize a :class:`FArg`.
-
-        :arg arg: a supertype of :class:`FArg`, from which this Arg is derived.
-        :arg gather: recognized values: ``postponed``, ``onlymap``. With ``postponed``,
-            the gather is performed at some in a callee of the wrapper function; with
-            ``onlymap``, the gather is performed as usual in the wrapper, but only
-            the map values are staged.
-        :arg c_index: if True, will provide the kernel with the iteration index of this
-            Arg's set. Otherwise, code generation is unaffected.
-        """
-        super(FArg, self).__init__(arg.data, arg.map, arg.idx, arg.access, arg._flatten)
-        self.position = arg.position
-        self.indirect_position = arg.indirect_position
-        self.gather = gather or arg.gather
-        self.c_index = c_index or arg.c_index
-
-        if hasattr(arg, 'hackflatten'):
-            self.hackflatten = True
-
-    def c_map_name(self, i, j, fromvector=False):
-        map_name = super(FArg, self).c_map_name(i, j)
-        return map_name if not fromvector else "&%s[0]" % map_name
-
-    def c_vec_dec(self, is_facet=False):
-        if self.gather == 'onlymap':
-            facet_mult = 2 if is_facet else 1
-            cdim = self.data.cdim if self._flatten else 1
-            return "%(type)s %(vec_name)s[%(arity)s];\n" % \
-                {'type': self.ctype,
-                 'vec_name': self.c_vec_name(),
-                 'arity': self.map.arity * cdim * facet_mult}
-        else:
-            return super(FArg, self).c_vec_dec(is_facet)
-
-    def c_vec_init(self, is_top, is_facet=False, force_gather=False):
-        if self.gather == 'postponed' and not force_gather:
-            return ''
-        elif self.gather == 'onlymap':
-            vec_name = self.c_vec_name()
-            map_name = self.c_map_name(0, 0)
-            arity = self.map.arity
-            return ';\n'.join(["%s[%s] = %s[%s*%s+%s]" %
-                               (vec_name, i, map_name, self.c_def_index(), arity, i)
-                               for i in range(self.map.arity)])
-        else:
-            return super(FArg, self).c_vec_init(is_top, is_facet)
-
-    def c_kernel_arg(self, count, i=0, j=0, shape=(0,), layers=1):
-        if self.gather == 'postponed':
-            c_args = "%s, %s" % (self.c_arg_name(i),
-                                 self.c_map_name(i, 0, self.c_map_is_vector()))
-        elif self.gather == 'onlymap':
-            c_args = "%s, %s" % (self.c_arg_name(i), self.c_vec_name())
-        else:
-            c_args = super(FArg, self).c_kernel_arg(count, i, j, shape, layers)
-        if self.c_index:
-            c_args += ", %s" % self.c_def_index()
-        return c_args
-
-    def c_def_index(self):
-        return 'i'
-
-    def c_map_is_vector(self):
-        return False
-
-
-class TileArg(FArg):
-
-    """An Arg specialized for kernels and loops subjected to tiling."""
-
-    def __init__(self, arg, loop_position, gtl_maps=None):
-        """Initialize a :class:`TileArg`.
-
-        :arg arg: a supertype of :class:`TileArg`, from which this Arg is derived.
-        :arg loop_position: the position of the loop in the loop chain that this
-            object belongs to.
-        :arg gtl_maps: a dict associating global map names to local map names.
-        """
-        super(TileArg, self).__init__(arg)
-        self.loop_position = loop_position
-
-        c_local_maps = None
-        maps = as_tuple(arg.map, Map)
-        if gtl_maps:
-            c_local_maps = [None]*len(maps)
-            for i, map in enumerate(maps):
-                c_local_maps[i] = [None]*len(map)
-                for j, m in enumerate(map):
-                    c_local_maps[i][j] = gtl_maps["%s%d_%d" % (m.name, i, j)]
-        self._c_local_maps = c_local_maps
-
-    def c_arg_bindto(self):
-        """Assign this Arg's c_pointer to ``arg``."""
-        return "%s* %s = %s" % (self.ctype, self.c_arg_name(), self.ref_arg.c_arg_name())
-
-    def c_ind_data(self, idx, i, j=0, is_top=False, offset=None, var=None):
-        if not var:
-            var = 'i' if not self._c_local_maps else 'n'
-        return super(TileArg, self).c_ind_data(idx, i, j, is_top, offset, var)
-
-    def c_map_name(self, i, j, fromvector=False):
-        if not self._c_local_maps:
-            map_name = host.Arg.c_map_name(self.ref_arg, i, j)
-        else:
-            map_name = self._c_local_maps[i][j]
-        return map_name if not fromvector else "&%s[0]" % map_name
-
-    def c_map_entry(self, var):
-        maps = []
-        for idx in range(self.map.arity):
-            maps.append("%(map_name)s[%(var)s * %(arity)d + %(idx)d]" % {
-                'map_name': self.c_map_name(0, 0),
-                'var': var,
-                'arity': self.map.arity,
-                'idx': idx
-            })
-        return maps
-
-    def c_vec_entry(self, var, only_base=False):
-        vecs = []
-        for idx in range(self.map.arity):
-            for k in range(self.data.cdim):
-                vecs.append(self.c_ind_data(idx, 0, k, var=var))
-                if only_base:
-                    break
-        return vecs
-
-    def c_global_reduction_name(self, count=None):
-        return "%(name)s_l%(count)d[0]" % {
-            'name': self.c_arg_name(),
-            'count': count}
-
-    def c_def_index(self):
-        return 'i' if not self._c_local_maps else 'n'
-
-    def c_map_is_vector(self):
-        return False if not self._c_local_maps else True
-
-    @property
-    def name(self):
-        """The generated argument name."""
-        return "arg_exec_loop%d_%d" % (self.loop_position, self.position)
-
-
-class Kernel(sequential.Kernel, tuple):
-
-    """A :class:`fusion.Kernel` represents a sequence of kernels.
-
-    The sequence can be:
-
-        * the result of the concatenation of kernel bodies (so a single C function
-            is present)
-        * a list of separate kernels (multiple C functions, which have to be
-            suitably called within the wrapper function)."""
-
-    @classmethod
-    def _cache_key(cls, kernels, fused_ast=None, loop_chain_index=None):
-        keys = "".join([super(Kernel, cls)._cache_key(
-            k._original_ast.gencode() if k._original_ast else k._code,
-            k._name, k._opts, k._include_dirs, k._headers, k._user_code) for k in kernels])
-        return str(loop_chain_index) + keys
-
-    def _ast_to_c(self, asts, opts):
-        """Produce a string of C code from an abstract syntax tree representation
-        of the kernel."""
-        if not isinstance(asts, (ast.FunDecl, ast.Root)):
-            asts = ast.Root(asts)
-        self._ast = asts
-        self._original_ast = dcopy(self._ast)
-        return super(Kernel, self)._ast_to_c(self._ast, opts)
-
-    def _multiple_ast_to_c(self, kernels):
-        """Glue together different ASTs (or strings) such that: ::
-
-            * clashes due to identical function names are avoided;
-            * duplicate functions (same name, same body) are avoided.
-        """
-        code = ""
-        identifier = lambda k: k.cache_key[1:]
-        unsorted_kernels = sorted(kernels, key=identifier)
-        for i, (_, kernel_group) in enumerate(groupby(unsorted_kernels, identifier)):
-            duplicates = list(kernel_group)
-            main = duplicates[0]
-            if main._original_ast:
-                main_ast = dcopy(main._original_ast)
-                finder = FindInstances((ast.FunDecl, ast.FunCall))
-                found = finder.visit(main_ast, ret=FindInstances.default_retval())
-                for fundecl in found[ast.FunDecl]:
-                    new_name = "%s_%d" % (fundecl.name, i)
-                    # Need to change the name of any inner functions too
-                    for funcall in found[ast.FunCall]:
-                        if fundecl.name == funcall.funcall.symbol:
-                            funcall.funcall.symbol = new_name
-                    fundecl.name = new_name
-                function_name = "%s_%d" % (main._name, i)
-                code += host.Kernel._ast_to_c(main, main_ast, main._opts)
-            else:
-                # AST not available so can't change the name, hopefully there
-                # will not be compile time clashes.
-                function_name = main._name
-                code += main._code
-            # Finally track the function name within this /fusion.Kernel/
-            for k in duplicates:
-                try:
-                    k._function_names[self.cache_key] = function_name
-                except AttributeError:
-                    k._function_names = {
-                        k.cache_key: k.name,
-                        self.cache_key: function_name
-                    }
-            code += "\n"
-        return code
-
-    def __init__(self, kernels, fused_ast=None, loop_chain_index=None):
-        """Initialize a :class:`fusion.Kernel` object.
-
-        :arg kernels: an iterator of some :class:`Kernel` objects. The objects
-            can be of class `fusion.Kernel` or of any superclass.
-        :arg fused_ast: the abstract syntax tree of the fused kernel. If not
-            provided, objects in ``kernels`` are considered "isolated C functions".
-        :arg loop_chain_index: index (i.e., position) of the kernel in a loop chain.
-            Meaningful only if ``fused_ast`` is specified.
-        """
-        # Protect against re-initialization when retrieved from cache
-        if self._initialized:
-            return
-        Kernel._globalcount += 1
-
-        # We need to distinguish between the kernel name and the function name(s).
-        # Since /fusion.Kernel/ are, in general, collections of functions, the same
-        # function (which is itself associated a Kernel) can appear in different
-        # /fusion.Kernel/ objects, but possibly under a different name (to avoid
-        # name clashes)
-        self._name = "_".join([k.name for k in kernels])
-        self._function_names = {self.cache_key: self._name}
-
-        self._cpp = any(k._cpp for k in kernels)
-        self._opts = dict(flatten([k._opts.items() for k in kernels]))
-        self._applied_blas = any(k._applied_blas for k in kernels)
-        self._include_dirs = list(set(flatten([k._include_dirs for k in kernels])))
-        self._ldargs = list(set(flatten([k._ldargs for k in kernels])))
-        self._headers = list(set(flatten([k._headers for k in kernels])))
-        self._user_code = "\n".join(list(set([k._user_code for k in kernels])))
-        self._attached_info = False
-
-        # What sort of Kernel do I have?
-        if fused_ast:
-            # A single, already fused AST (code generation is then delayed)
-            self._ast = fused_ast
-            self._code = None
-        else:
-            # Multiple kernels, interpreted as different C functions
-            self._ast = None
-            self._code = self._multiple_ast_to_c(kernels)
-        self._original_ast = self._ast
-        self._kernels = kernels
-
-        self._initialized = True
-
-    def __iter__(self):
-        for k in self._kernels:
-            yield k
-
-    def __str__(self):
-        return "OP2 FusionKernel: %s" % self._name
-
-
-# Parallel loop API
-
-class IterationSpace(base.IterationSpace):
-
-    """A simple bag of :class:`IterationSpace` objects."""
-
-    def __init__(self, all_itspaces):
-        self._iterset = [i._iterset for i in all_itspaces]
-
-    def __str__(self):
-        output = "OP2 Fused Iteration Space:"
-        output += "\n  ".join(["%s with extents %s" % (i._iterset, i._extents)
-                               for i in self.iterset])
-        return output
-
-    def __repr__(self):
-        return "\n".join(["IterationSpace(%r, %r)" % (i._iterset, i._extents)
-                          for i in self.iterset])
-
-
-class JITModule(sequential.JITModule):
-
-    _cppargs = []
-    _libraries = []
-    _extension = 'cpp'
-
-    _wrapper = """
-extern "C" void %(wrapper_name)s(%(executor_arg)s,
-                      %(ssinds_arg)s
-                      %(wrapper_args)s
-                      %(const_args)s
-                      %(rank)s
-                      %(region_flag)s);
-void %(wrapper_name)s(%(executor_arg)s,
-                      %(ssinds_arg)s
-                      %(wrapper_args)s
-                      %(const_args)s
-                      %(rank)s
-                      %(region_flag)s) {
-  %(user_code)s
-  %(wrapper_decs)s;
-  %(const_inits)s;
-
-  %(executor_code)s;
-}
-"""
-    _kernel_wrapper = """
-%(interm_globals_decl)s;
-%(interm_globals_init)s;
-%(vec_decs)s;
-%(args_binding)s;
-%(tile_init)s;
-for (int n = %(tile_start)s; n < %(tile_end)s; n++) {
-  int i = %(tile_iter)s;
-  %(prefetch_maps)s;
-  %(vec_inits)s;
-  %(prefetch_vecs)s;
-  %(buffer_decl)s;
-  %(buffer_gather)s
-  %(kernel_name)s(%(kernel_args)s);
-  i = %(index_expr)s;
-  %(itset_loop_body)s;
-}
-%(tile_finish)s;
-%(interm_globals_writeback)s;
-"""
-
-    @classmethod
-    def _cache_key(cls, kernel, itspace, *args, **kwargs):
-        insp_name = kwargs['insp_name']
-        key = (insp_name, kwargs['use_glb_maps'], kwargs['use_prefetch'])
-        if insp_name != lazy_trace_name:
-            return key
-        all_kernels = kwargs['all_kernels']
-        all_itspaces = kwargs['all_itspaces']
-        all_args = kwargs['all_args']
-        for kernel, itspace, args in zip(all_kernels, all_itspaces, all_args):
-            key += super(JITModule, cls)._cache_key(kernel, itspace, *args)
-        return key
-
-    def __init__(self, kernel, itspace, *args, **kwargs):
-        if self._initialized:
-            return
-        self._all_kernels = kwargs.pop('all_kernels')
-        self._all_itspaces = kwargs.pop('all_itspaces')
-        self._all_args = kwargs.pop('all_args')
-        self._executor = kwargs.pop('executor')
-        self._use_glb_maps = kwargs.pop('use_glb_maps')
-        self._use_prefetch = kwargs.pop('use_prefetch')
-        super(JITModule, self).__init__(kernel, itspace, *args, **kwargs)
-
-    def set_argtypes(self, iterset, *args):
-        argtypes = [slope.Executor.meta['py_ctype_exec']]
-        for itspace in self._all_itspaces:
-            if isinstance(itspace.iterset, Subset):
-                argtypes.append(itspace.iterset._argtype)
-        for arg in args:
-            if arg._is_mat:
-                argtypes.append(arg.data._argtype)
-            else:
-                for d in arg.data:
-                    argtypes.append(d._argtype)
-            if arg._is_indirect or arg._is_mat:
-                maps = as_tuple(arg.map, Map)
-                for map in maps:
-                    for m in map:
-                        argtypes.append(m._argtype)
-        for c in Const._definitions():
-            argtypes.append(c._argtype)
-
-        # MPI related stuff (rank, region)
-        argtypes.append(ctypes.c_int)
-        argtypes.append(ctypes.c_int)
-
-        self._argtypes = argtypes
-
-    def compile(self):
-        # If we weren't in the cache we /must/ have arguments
-        if not hasattr(self, '_args'):
-            raise RuntimeError("JITModule not in cache, but has no args associated")
-
-        # Set compiler and linker options
-        slope_dir = os.environ['SLOPE_DIR']
-        self._kernel._name = 'executor'
-        self._kernel._headers.extend(slope.Executor.meta['headers'])
-        if self._use_prefetch:
-            self._kernel._headers.extend(['#include "xmmintrin.h"'])
-        self._kernel._include_dirs.extend(['%s/%s' % (slope_dir,
-                                                      slope.get_include_dir())])
-        self._libraries += ['-L%s/%s' % (slope_dir, slope.get_lib_dir()),
-                            '-l%s' % slope.get_lib_name()]
-        compiler = coffee.plan.compiler.get('name')
-        self._cppargs += slope.get_compile_opts(compiler)
-        fun = super(JITModule, self).compile()
-
-        if hasattr(self, '_all_args'):
-            # After the JITModule is compiled, can drop any reference to now
-            # useless fields
-            del self._all_kernels
-            del self._all_itspaces
-            del self._all_args
-            del self._executor
-
-        return fun
-
-    def generate_code(self):
-        indent = lambda t, i: ('\n' + '  ' * i).join(t.split('\n'))
-
-        # 1) Construct the wrapper arguments
-        code_dict = {}
-        code_dict['wrapper_name'] = 'wrap_executor'
-        code_dict['executor_arg'] = "%s %s" % (slope.Executor.meta['ctype_exec'],
-                                               slope.Executor.meta['name_param_exec'])
-        _wrapper_args = ', '.join([arg.c_wrapper_arg() for arg in self._args])
-        _wrapper_decs = ';\n'.join([arg.c_wrapper_dec() for arg in self._args])
-        code_dict['wrapper_args'] = _wrapper_args
-        code_dict['wrapper_decs'] = indent(_wrapper_decs, 1)
-        code_dict['rank'] = ", %s %s" % (slope.Executor.meta['ctype_rank'],
-                                         slope.Executor.meta['rank'])
-        code_dict['region_flag'] = ", %s %s" % (slope.Executor.meta['ctype_region_flag'],
-                                                slope.Executor.meta['region_flag'])
-
-        # 2) Construct the kernel invocations
-        _loop_body, _user_code, _ssinds_arg = [], [], []
-        _const_args, _const_inits = set(), set()
-        # For each kernel ...
-        for i, (kernel, it_space, args) in enumerate(zip(self._all_kernels,
-                                                         self._all_itspaces,
-                                                         self._all_args)):
-            # ... bind the Executor's arguments to this kernel's arguments
-            binding = []
-            for a1 in args:
-                for a2 in self._args:
-                    if a1.data is a2.data and a1.map is a2.map:
-                        a1.ref_arg = a2
-                        break
-                binding.append(a1.c_arg_bindto())
-            binding = ";\n".join(binding)
-
-            # ... obtain the /code_dict/ as if it were not part of an Executor,
-            # since bits of code generation can be reused
-            loop_code_dict = sequential.JITModule(kernel, it_space, *args, delay=True)
-            loop_code_dict = loop_code_dict.generate_code()
-
-            # ... does the scatter use global or local maps ?
-            if self._use_glb_maps:
-                loop_code_dict['index_expr'] = '%s[n]' % self._executor.gtl_maps[i]['DIRECT']
-                prefetch_var = 'int p = %s[n + %d]' % (self._executor.gtl_maps[i]['DIRECT'],
-                                                        self._use_prefetch)
-            else:
-                prefetch_var = 'int p = n + %d' % self._use_prefetch
-
-            # ... add prefetch intrinsics, if requested
-            prefetch_maps, prefetch_vecs = '', ''
-            if self._use_prefetch:
-                prefetch = lambda addr: '_mm_prefetch ((char*)(%s), _MM_HINT_T0)' % addr
-                prefetch_maps = [a.c_map_entry('p') for a in args if a._is_indirect]
-                # can save some instructions since prefetching targets chunks of 32 bytes
-                prefetch_maps = flatten([j for j in pm if pm.index(j) % 2 == 0]
-                                        for pm in prefetch_maps)
-                prefetch_maps = list(OrderedDict.fromkeys(prefetch_maps))
-                prefetch_maps = ';\n'.join([prefetch_var] +
-                                           [prefetch('&(%s)' % pm) for pm in prefetch_maps])
-                prefetch_vecs = flatten(a.c_vec_entry('p', True) for a in args
-                                        if a._is_indirect)
-                prefetch_vecs = ';\n'.join([prefetch(pv) for pv in prefetch_vecs])
-            loop_code_dict['prefetch_maps'] = prefetch_maps
-            loop_code_dict['prefetch_vecs'] = prefetch_vecs
-
-            # ... build the subset indirection array, if necessary
-            _ssind_arg, _ssind_decl = '', ''
-            if loop_code_dict['ssinds_arg']:
-                _ssind_arg = 'ssinds_%d' % i
-                _ssind_decl = 'int* %s' % _ssind_arg
-                loop_code_dict['index_expr'] = '%s[n]' % _ssind_arg
-
-            # ... use the proper function name (the function name of the kernel
-            # within *this* specific loop chain)
-            loop_code_dict['kernel_name'] = kernel._function_names[self._kernel.cache_key]
-
-            # ... finish building up the /code_dict/
-            loop_code_dict['args_binding'] = binding
-            loop_code_dict['tile_init'] = self._executor.c_loop_init[i]
-            loop_code_dict['tile_finish'] = self._executor.c_loop_end[i]
-            loop_code_dict['tile_start'] = slope.Executor.meta['tile_start']
-            loop_code_dict['tile_end'] = slope.Executor.meta['tile_end']
-            loop_code_dict['tile_iter'] = '%s[n]' % self._executor.gtl_maps[i]['DIRECT']
-            if _ssind_arg:
-                loop_code_dict['tile_iter'] = '%s[%s]' % (_ssind_arg, loop_code_dict['tile_iter'])
-
-            # ... concatenate the rest, i.e., body, user code, constants, ...
-            _loop_body.append(strip(JITModule._kernel_wrapper % loop_code_dict))
-            _user_code.append(kernel._user_code)
-            _ssinds_arg.append(_ssind_decl)
-            _const_args.add(loop_code_dict['const_args'])
-            _const_inits.add(loop_code_dict['const_inits'])
-
-        _loop_chain_body = indent("\n\n".join(_loop_body), 2)
-        code_dict['const_args'] = "".join(_const_args)
-        code_dict['const_inits'] = indent("".join(_const_inits), 1)
-        code_dict['user_code'] = indent("\n".join(_user_code), 1)
-        code_dict['ssinds_arg'] = "".join(["%s," % s for s in _ssinds_arg if s])
-        code_dict['executor_code'] = indent(self._executor.c_code(_loop_chain_body), 1)
-
-        return code_dict
-
-
-class ParLoop(sequential.ParLoop):
-
-    def __init__(self, kernel, it_space, *args, **kwargs):
-        LazyComputation.__init__(self,
-                                 kwargs['read_args'] | Const._defs,
-                                 kwargs['written_args'],
-                                 kwargs['inc_args'])
-
-        # Inspector related stuff
-        self._all_kernels = kwargs.get('all_kernels', [kernel])
-        self._all_itspaces = kwargs.get('all_itspaces', [kernel])
-        self._all_args = kwargs.get('all_args', [args])
-        self._insp_name = kwargs.get('insp_name')
-        self._inspection = kwargs.get('inspection')
-        # Executor related stuff
-        self._executor = kwargs.get('executor')
-        self._use_glb_maps = kwargs.get('use_glb_maps')
-        self._use_prefetch = kwargs.get('use_prefetch')
-
-        # Global reductions are obviously forbidden when tiling; however, the user
-        # might have bypassed this condition because sure about safety. Therefore,
-        # we act as in the super class, computing the result in a temporary buffer,
-        # and then copying it back into the original input. This is for safety of
-        # parallel global reductions (for more details, see base.ParLoop)
-        self._reduced_globals = {}
-        for _globs, _args in zip(kwargs.get('reduced_globals', []), self._all_args):
-            if not _globs:
-                continue
-            for i, glob in _globs.iteritems():
-                shadow_glob = _args[i].data
-                for j, data in enumerate([a.data for a in args]):
-                    if shadow_glob is data:
-                        self._reduced_globals[j] = glob
-                        break
-
-        self._kernel = kernel
-        self._actual_args = args
-        self._it_space = it_space
-        self._only_local = False
-
-        for i, arg in enumerate(self._actual_args):
-            arg.name = "arg%d" % i  # Override the previously cached_property name
-            arg.position = i
-            arg.indirect_position = i
-        for i, arg1 in enumerate(self._actual_args):
-            if arg1._is_dat and arg1._is_indirect:
-                for arg2 in self._actual_args[i:]:
-                    # We have to check for identity here (we really
-                    # want these to be the same thing, not just look
-                    # the same)
-                    if arg2.data is arg1.data and arg2.map is arg1.map:
-                        arg2.indirect_position = arg1.indirect_position
-
-    def prepare_arglist(self, part, *args):
-        arglist = [self._inspection]
-        for itspace in self._all_itspaces:
-            if isinstance(itspace._iterset, Subset):
-                arglist.append(itspace._iterset._indices.ctypes.data)
-        for arg in args:
-            if arg._is_mat:
-                arglist.append(arg.data.handle.handle)
-            else:
-                for d in arg.data:
-                    # Cannot access a property of the Dat or we will force
-                    # evaluation of the trace
-                    arglist.append(d._data.ctypes.data)
-
-            if arg._is_indirect or arg._is_mat:
-                maps = as_tuple(arg.map, Map)
-                for map in maps:
-                    for m in map:
-                        arglist.append(m._values.ctypes.data)
-
-        for c in Const._definitions():
-            arglist.append(c._data.ctypes.data)
-
-        arglist.append(MPI.comm.rank)
-
-        return arglist
-
-    @collective
-    def compute(self):
-        """Execute the kernel over all members of the iteration space."""
-        with timed_region("ParLoopChain: executor (%s)" % self._insp_name):
-            self.halo_exchange_begin()
-            kwargs = {
-                'all_kernels': self._all_kernels,
-                'all_itspaces': self._all_itspaces,
-                'all_args': self._all_args,
-                'executor': self._executor,
-                'insp_name': self._insp_name,
-                'use_glb_maps': self._use_glb_maps,
-                'use_prefetch': self._use_prefetch
-            }
-            fun = JITModule(self.kernel, self.it_space, *self.args, **kwargs)
-            arglist = self.prepare_arglist(None, *self.args)
-            fun(*(arglist + [0]))
-            self.halo_exchange_end()
-            fun(*(arglist + [1]))
-            # Only meaningful if the user is enforcing tiling in presence of
-            # global reductions
-            self.reduction_begin()
-            self.reduction_end()
-            self.update_arg_data_state()
-
-
-# Utility classes
-
-class Filter(object):
-
-    """A utility class for filtering arguments originating from a set of
-    parallel loops. Arguments are filtered based on the data they contain
-    as well as the map used for accessing the data."""
-
-    def _key(self, arg):
-        return (arg.data, arg.map)
-
-    def loop_args(self, loops):
-        loop_args = [loop.args for loop in loops]
-        filtered_args = OrderedDict()
-        for args in loop_args:
-            for a in args:
-                fa = filtered_args.setdefault(self._key(a), a)
-                if a.access != fa.access:
-                    if READ in [a.access, fa.access]:
-                        # If a READ and some sort of write (MIN, MAX, RW, WRITE,
-                        # INC), then the access mode becomes RW
-                        fa.access = RW
-                    elif WRITE in [a.access, fa.access]:
-                        # Can't be a READ, so just stick to WRITE regardless of what
-                        # the other access mode is
-                        fa.access = WRITE
-                    else:
-                        # Neither READ nor WRITE, so access modes are some
-                        # combinations of RW, INC, MIN, MAX. For simplicity,
-                        # just make it RW.
-                        fa.access = RW
-        return filtered_args
-
-    def kernel_args(self, loops, fundecl):
-        """Eliminate redundant arguments in the kernel signature."""
-        loop_args = list(flatten([l.args for l in loops]))
-        unique_loop_args = self.loop_args(loops)
-        kernel_args = fundecl.args
-        binding = OrderedDict(zip(loop_args, kernel_args))
-        new_kernel_args, args_maps = [], []
-        for loop_arg, kernel_arg in binding.items():
-            key = self._key(loop_arg)
-            unique_loop_arg = unique_loop_args[key]
-            if loop_arg is unique_loop_arg:
-                new_kernel_args.append(kernel_arg)
-                continue
-            tobind_kernel_arg = binding[unique_loop_arg]
-            if tobind_kernel_arg.is_const:
-                # Need to remove the /const/ qualifier from the C declaration
-                # if the same argument is written to, somewhere, in the kernel.
-                # Otherwise, /const/ must be appended, if not present already,
-                # to the alias' qualifiers
-                if loop_arg._is_written:
-                    tobind_kernel_arg.qual.remove('const')
-                elif 'const' not in kernel_arg.qual:
-                    kernel_arg.qual.append('const')
-            # Update the /binding/, since might be useful for the caller
-            binding[loop_arg] = tobind_kernel_arg
-            # Aliases may be created instead of changing symbol names
-            if kernel_arg.sym.symbol == tobind_kernel_arg.sym.symbol:
-                continue
-            alias = ast_make_alias(dcopy(kernel_arg), dcopy(tobind_kernel_arg))
-            args_maps.append(alias)
-        fundecl.children[0].children = args_maps + fundecl.children[0].children
-        fundecl.args = new_kernel_args
-        return binding
-
-
-class WeakFilter(Filter):
-
-    """Filter arguments based on the data they contain."""
-
-    def _key(self, arg):
-        return arg.data
-
-
-# An Inspector produces one of the following Schedules
-
-class Schedule(object):
-
-    """Represent an execution scheme for a sequence of :class:`ParLoop` objects."""
-
-    def __init__(self, insp_name, schedule=None):
-        self._insp_name = insp_name
-        self._schedule = schedule
-
-    def __call__(self, loop_chain):
-        """Given an iterator of :class:`ParLoop` objects (``loop_chain``),
-        return an iterator of new :class:`ParLoop` objects. The input parloops
-        are "scheduled" according to the strategy of this Schedule. The Schedule
-        itself was produced by an Inspector.
-
-        In the simplest case, the returned value is identical to the input
-        ``loop_chain``. That is, the Inspector that created this Schedule could
-        not apply any fusion or tiling.
-
-        In general, the Schedule could fuse or tile the loops in ``loop_chain``.
-        A sequence of  :class:`fusion.ParLoop` objects would then be returned.
-        """
-        return loop_chain
-
-    def _filter(self, loops):
-        return Filter().loop_args(loops).values()
-
-
-class PlainSchedule(Schedule):
-
-    def __init__(self, insp_name, kernels):
-        super(PlainSchedule, self).__init__(insp_name)
-        self._kernel = kernels
-
-    def __call__(self, loop_chain):
-        for loop in loop_chain:
-            for arg in loop.args:
-                arg.gather = None
-                arg.c_index = False
-        return loop_chain
-
-
-class FusionSchedule(Schedule):
-
-    """Schedule an iterator of :class:`ParLoop` objects applying soft fusion."""
-
-    def __init__(self, insp_name, schedule, kernels, offsets):
-        super(FusionSchedule, self).__init__(insp_name, schedule)
-        self._kernel = list(kernels)
-
-        # Track the /ParLoop/s in the loop chain that each fused kernel maps to
-        offsets = [0] + list(offsets)
-        loop_indices = [range(offsets[i], o) for i, o in enumerate(offsets[1:])]
-        self._info = [{'loop_indices': li} for li in loop_indices]
-
-    def __call__(self, loop_chain):
-        loop_chain = self._schedule(loop_chain)
-        fused_par_loops = []
-        for kernel, info in zip(self._kernel, self._info):
-            loop_indices = info['loop_indices']
-            extra_args = info.get('extra_args', [])
-            # Create the ParLoop's arguments. Note that both the iteration set and
-            # the iteration region must correspond to that of the /base/ loop
-            iterregion = loop_chain[loop_indices[0]].iteration_region
-            iterset = loop_chain[loop_indices[0]].it_space.iterset
-            args = self._filter([loop_chain[i] for i in loop_indices])
-            # Create any ParLoop additional arguments
-            extra_args = [Dat(*d)(*a) for d, a in extra_args]
-            args += extra_args
-            # Remove now incorrect cached properties:
-            for a in args:
-                a.__dict__.pop('name', None)
-            # Create the actual ParLoop, resulting from the fusion of some kernels
-            fused_par_loops.append(_make_object('ParLoop', kernel, iterset, *args,
-                                                iterate=iterregion,
-                                                insp_name=self._insp_name))
-        return fused_par_loops
-
-
-class HardFusionSchedule(FusionSchedule, Schedule):
-
-    """Schedule an iterator of :class:`ParLoop` objects applying hard fusion
-    on top of soft fusion."""
-
-    def __init__(self, insp_name, schedule, fused):
-        Schedule.__init__(self, insp_name, schedule)
-        self._fused = fused
-
-        # Set proper loop_indices for this schedule
-        self._info = dcopy(schedule._info)
-        for i, info in enumerate(schedule._info):
-            for k, v in info.items():
-                self._info[i][k] = [i] if k == 'loop_indices' else v
-
-        # Update the input schedule to make use of hard fusion kernels
-        kernel = scopy(schedule._kernel)
-        for ofs, (fused_kernel, fused_map, fargs) in enumerate(fused):
-            # Find the position of the /fused/ kernel in the new loop chain.
-            base, fuse = fused_kernel._kernels
-            base_idx, fuse_idx = kernel.index(base), kernel.index(fuse)
-            pos = min(base_idx, fuse_idx)
-            self._info[pos]['loop_indices'] = [base_idx + ofs, fuse_idx + ofs]
-            # A bitmap indicates whether the i-th iteration in /fuse/ has been executed
-            self._info[pos]['extra_args'] = [((fused_map.toset, None, np.int32),
-                                              (RW, fused_map))]
-            # Keep track of the arguments needing a postponed gather
-            self._info[pos]['fargs'] = fargs
-            # Now we can modify the kernel sequence
-            kernel.insert(pos, fused_kernel)
-            kernel.pop(pos+1)
-            pos = max(base_idx, fuse_idx)
-            self._info.pop(pos)
-            kernel.pop(pos)
-        self._kernel = kernel
-
-    def __call__(self, loop_chain, only_hard=False):
-        # First apply soft fusion, then hard fusion
-        if not only_hard:
-            loop_chain = self._schedule(loop_chain)
-        fused_par_loops = FusionSchedule.__call__(self, loop_chain)
-        for i, (loop, info) in enumerate(zip(list(fused_par_loops), self._info)):
-            fargs = info.get('fargs', {})
-            args = [FArg(arg, *fargs[j]) if j in fargs else arg
-                    for j, arg in enumerate(loop.args)]
-            fused_par_loop = _make_object('ParLoop', loop.kernel, loop.it_space.iterset,
-                                          *tuple(args), iterate=loop.iteration_region,
-                                          insp_name=self._insp_name)
-            fused_par_loops[i] = fused_par_loop
-        return fused_par_loops
-
-    def _filter(self, loops):
-        return WeakFilter().loop_args(loops).values()
-
-
-class TilingSchedule(Schedule):
-
-    """Schedule an iterator of :class:`ParLoop` objects applying tiling, possibly on
-    top of hard fusion and soft fusion."""
-
-    def __init__(self, insp_name, schedule, kernel, inspection, executor, **options):
-        super(TilingSchedule, self).__init__(insp_name, schedule)
-        self._inspection = inspection
-        self._executor = executor
-        self._kernel = kernel
-        # Schedule's optimizations
-        self._opt_glb_maps = options.get('use_glb_maps', False)
-        self._opt_prefetch = options.get('use_prefetch', 0)
-
-    def __call__(self, loop_chain):
-        loop_chain = self._schedule(loop_chain)
-        # Track the individual kernels, and the args of each kernel
-        all_itspaces = tuple(loop.it_space for loop in loop_chain)
-        all_args = []
-        for i, (loop, gtl_maps) in enumerate(zip(loop_chain, self._executor.gtl_maps)):
-            all_args.append([TileArg(arg, i, None if self._opt_glb_maps else gtl_maps)
-                             for arg in loop.args])
-        all_args = tuple(all_args)
-        # Data for the actual ParLoop
-        it_space = IterationSpace(all_itspaces)
-        args = self._filter(loop_chain)
-        reduced_globals = [loop._reduced_globals for loop in loop_chain]
-        read_args = set(flatten([loop.reads for loop in loop_chain]))
-        written_args = set(flatten([loop.writes for loop in loop_chain]))
-        inc_args = set(flatten([loop.incs for loop in loop_chain]))
-        kwargs = {
-            'all_kernels': self._kernel._kernels,
-            'all_itspaces': all_itspaces,
-            'all_args': all_args,
-            'read_args': read_args,
-            'written_args': written_args,
-            'reduced_globals': reduced_globals,
-            'inc_args': inc_args,
-            'insp_name': self._insp_name,
-            'use_glb_maps': self._opt_glb_maps,
-            'use_prefetch': self._opt_prefetch,
-            'inspection': self._inspection,
-            'executor': self._executor
-        }
-        return [ParLoop(self._kernel, it_space, *args, **kwargs)]
-
-
-# Loop chain inspection
-
-class Inspector(Cached):
-
-    """An Inspector constructs a Schedule to fuse or tile a sequence of loops.
-
-    .. note:: For tiling, the Inspector relies on the SLOPE library."""
-
-    _cache = {}
-    _modes = ['soft', 'hard', 'tile', 'only_tile', 'only_omp']
-
-    @classmethod
-    def _cache_key(cls, name, loop_chain, **options):
-        key = (name,)
-        if name != lazy_trace_name:
-            # Special case: the Inspector comes from a user-defined /loop_chain/
-            key += (options['mode'], options['tile_size'],
-                    options['use_glb_maps'], options['use_prefetch'], options['coloring'])
-            key += (loop_chain[0].kernel.cache_key,)
-            return key
-        # Inspector extracted from lazy evaluation trace
-        for loop in loop_chain:
-            if isinstance(loop, base._LazyMatOp):
-                continue
-            key += (loop.kernel.cache_key,)
-            key += (loop.it_space.cache_key, loop.it_space.iterset.sizes)
-            for arg in loop.args:
-                if arg._is_global:
-                    key += (arg.data.dim, arg.data.dtype, arg.access)
-                elif arg._is_dat:
-                    if isinstance(arg.idx, IterationIndex):
-                        idx = (arg.idx.__class__, arg.idx.index)
-                    else:
-                        idx = arg.idx
-                    map_arity = arg.map.arity if arg.map else None
-                    key += (arg.data.dim, arg.data.dtype, map_arity, idx, arg.access)
-                elif arg._is_mat:
-                    idxs = (arg.idx[0].__class__, arg.idx[0].index,
-                            arg.idx[1].index)
-                    map_arities = (arg.map[0].arity, arg.map[1].arity)
-                    key += (arg.data.dims, arg.data.dtype, idxs, map_arities, arg.access)
-        return key
-
-    def __init__(self, name, loop_chain, **options):
-        """Initialize an Inspector object.
-
-        :arg name: a name for the Inspector
-        :arg loop_chain: an iterator for the loops that will be fused/tiled
-        :arg options: a set of parameters to drive fusion/tiling
-            * mode: can take any of the values in ``Inspector._modes``, namely
-                soft, hard, tile, only_tile, only_omp:
-                * soft: consecutive loops over the same iteration set that do
-                    not present RAW or WAR dependencies through indirections
-                    are fused.
-                * hard: ``soft`` fusion; then, loops over different iteration sets
-                    are also fused, provided that there are no RAW or WAR
-                    dependencies.
-                * tile: ``soft`` and ``hard`` fusion; then, tiling through the
-                    SLOPE library takes place.
-                * only_tile: only tiling through the SLOPE library (i.e., no fusion)
-                * only_omp: ompize individual parloops through the SLOPE library
-            * tile_size: starting average tile size
-            * extra_halo: are we providing SLOPE with extra halo to be efficient
-                and allow it to minimize redundant computation ?
-        """
-        if self._initialized:
-            return
-        self._name = name
-        self._loop_chain = loop_chain
-        self._mode = options.pop('mode')
-        self._options = options
-        self._schedule = PlainSchedule(name, [loop.kernel for loop in self._loop_chain])
-
-    def inspect(self):
-        """Inspect the loop chain and produce a :class:`Schedule`."""
-        if self._initialized:
-            # An inspection plan is in cache.
-            return self._schedule
-        elif self._heuristic_skip_inspection():
-            # Not in cache, and too premature for running a potentially costly inspection
-            del self._name
-            del self._loop_chain
-            del self._mode
-            del self._options
-            return self._schedule
-
-        # Is `mode` legal ?
-        if self.mode not in Inspector._modes:
-            raise RuntimeError("Inspection accepts only %s fusion modes", Inspector._modes)
-
-        with timed_region("ParLoopChain `%s`: inspector" % self._name):
-            if self.mode in ['soft', 'hard', 'tile']:
-                self._soft_fuse()
-            if self.mode in ['hard', 'tile']:
-                self._hard_fuse()
-            if self.mode in ['tile', 'only_tile', 'only_omp']:
-                self._tile()
-
-        # A schedule has been computed. The Inspector is initialized and therefore
-        # retrievable from cache. We then blow away everything we don't need any more.
-        self._initialized = True
-        del self._name
-        del self._loop_chain
-        del self._mode
-        del self._options
-        return self._schedule
-
-    def _heuristic_skip_inspection(self):
-        """Decide, heuristically, whether to run an inspection or not.
-        If tiling is not requested, then inspection is performed.
-        If tiling is requested, then inspection is performed on the third
-        invocation. The fact that an inspection for the same loop chain
-        is requested multiple times suggests the parloops originate in a
-        time stepping loop. The cost of building tiles in SLOPE-land would
-        then be amortized over several iterations."""
-        self._ninsps = self._ninsps + 1 if hasattr(self, '_ninsps') else 1
-        if self.mode in ['tile', 'only_tile'] and self._ninsps < 3:
-            return True
-        return False
-
-    def _soft_fuse(self):
-        """Fuse consecutive loops over the same iteration set by concatenating
-        kernel bodies and creating new :class:`ParLoop` objects representing
-        the fused sequence.
-
-        The conditions under which two loops over the same iteration set can
-        be soft fused are:
-
-            * They are both direct, OR
-            * One is direct and the other indirect
-
-        This is detailed in the paper::
-
-            "Mesh Independent Loop Fusion for Unstructured Mesh Applications"
-
-        from C. Bertolli et al.
-        """
-
-        def fuse(self, loops, loop_chain_index):
-            # Naming convention: here, we are fusing ASTs in /fuse_asts/ within
-            # /base_ast/. Same convention will be used in the /hard_fuse/ method
-            kernels = [l.kernel for l in loops]
-            fuse_asts = [k._original_ast if k._code else k._ast for k in kernels]
-            # Fuse the actual kernels' bodies
-            base_ast = dcopy(fuse_asts[0])
-            retval = FindInstances.default_retval()
-            base_fundecl = FindInstances(ast.FunDecl).visit(base_ast, ret=retval)[ast.FunDecl]
-            if len(base_fundecl) != 1:
-                raise RuntimeError("Fusing kernels, but found unexpected AST")
-            base_fundecl = base_fundecl[0]
-            for unique_id, _fuse_ast in enumerate(fuse_asts[1:], 1):
-                fuse_ast = dcopy(_fuse_ast)
-                retval = FindInstances.default_retval()
-                fuse_fundecl = FindInstances(ast.FunDecl).visit(fuse_ast, ret=retval)[ast.FunDecl]
-                if len(fuse_fundecl) != 1:
-                    raise RuntimeError("Fusing kernels, but found unexpected AST")
-                fuse_fundecl = fuse_fundecl[0]
-                # 1) Extend function name
-                base_fundecl.name = "%s_%s" % (base_fundecl.name, fuse_fundecl.name)
-                # 2) Concatenate the arguments in the signature
-                base_fundecl.args.extend(fuse_fundecl.args)
-                # 3) Uniquify symbols identifiers
-                retval = SymbolReferences.default_retval()
-                fuse_symbols = SymbolReferences().visit(fuse_ast, ret=retval)
-                for decl in fuse_fundecl.args:
-                    for symbol, _ in fuse_symbols[decl.sym.symbol]:
-                        symbol.symbol = "%s_%d" % (symbol.symbol, unique_id)
-                # 4) Scope and concatenate bodies
-                base_fundecl.children[0] = ast.Block(
-                    [ast.Block(base_fundecl.children[0].children, open_scope=True),
-                     ast.FlatBlock("\n\n// Begin of fused kernel\n\n"),
-                     ast.Block(fuse_fundecl.children[0].children, open_scope=True)])
-            # Eliminate redundancies in the /fused/ kernel signature
-            Filter().kernel_args(loops, base_fundecl)
-            # Naming convention
-            fused_ast = base_ast
-            return Kernel(kernels, fused_ast, loop_chain_index)
-
-        fused, fusing = [], [self._loop_chain[0]]
-        for i, loop in enumerate(self._loop_chain[1:]):
-            base_loop = fusing[-1]
-            if base_loop.it_space != loop.it_space or \
-                    (base_loop.is_indirect and loop.is_indirect):
-                # Fusion not legal
-                fused.append((fuse(self, fusing, len(fused)), i+1))
-                fusing = [loop]
-            elif (base_loop.is_direct and loop.is_direct) or \
-                    (base_loop.is_direct and loop.is_indirect) or \
-                    (base_loop.is_indirect and loop.is_direct):
-                # This loop is fusible. Also, can speculative go on searching
-                # for other loops to fuse
-                fusing.append(loop)
-            else:
-                raise RuntimeError("Unexpected loop chain structure while fusing")
-        if fusing:
-            fused.append((fuse(self, fusing, len(fused)), len(self._loop_chain)))
-
-        fused_kernels, offsets = zip(*fused)
-        self._schedule = FusionSchedule(self._name, self._schedule, fused_kernels, offsets)
-        self._loop_chain = self._schedule(self._loop_chain)
-
-    def _hard_fuse(self):
-        """Fuse consecutive loops over different iteration sets that do not
-        present RAW, WAR or WAW dependencies. For examples, two loops like: ::
-
-            par_loop(kernel_1, it_space_1,
-                     dat_1_1(INC, ...),
-                     dat_1_2(READ, ...),
-                     ...)
-
-            par_loop(kernel_2, it_space_2,
-                     dat_2_1(INC, ...),
-                     dat_2_2(READ, ...),
-                     ...)
-
-        where ``dat_1_1 == dat_2_1`` and, possibly (but not necessarily),
-        ``it_space_1 != it_space_2``, can be hard fused. Note, in fact, that
-        the presence of ``INC`` does not imply a real WAR dependency, because
-        increments are associative."""
-
-        reads = lambda l: set([a.data for a in l.args if a.access in [READ, RW]])
-        writes = lambda l: set([a.data for a in l.args if a.access in [RW, WRITE, MIN, MAX]])
-        incs = lambda l: set([a.data for a in l.args if a.access in [INC]])
-
-        def has_raw_or_war(loop1, loop2):
-            # Note that INC after WRITE is a special case of RAW dependency since
-            # INC cannot take place before WRITE.
-            return reads(loop2) & writes(loop1) or writes(loop2) & reads(loop1) or \
-                incs(loop1) & (writes(loop2) - incs(loop2)) or \
-                incs(loop2) & (writes(loop1) - incs(loop1))
-
-        def has_iai(loop1, loop2):
-            return incs(loop1) & incs(loop2)
-
-        def fuse(base_loop, loop_chain, fused):
-            """Try to fuse one of the loops in ``loop_chain`` with ``base_loop``."""
-            for loop in loop_chain:
-                if has_raw_or_war(loop, base_loop):
-                    # Can't fuse across loops preseting RAW or WAR dependencies
-                    return []
-                if loop.it_space == base_loop.it_space:
-                    warning("Ignoring unexpected sequence of loops in loop fusion")
-                    continue
-                # Is there an overlap in any of the incremented regions? If that is
-                # the case, then fusion can really be beneficial
-                common_inc_data = has_iai(base_loop, loop)
-                if not common_inc_data:
-                    continue
-                common_incs = [a for a in base_loop.args + loop.args
-                               if a.data in common_inc_data]
-                # Hard fusion potentially doable provided that we own a map between
-                # the iteration spaces involved
-                maps = list(set(flatten([a.map for a in common_incs])))
-                maps += [m.factors for m in maps if hasattr(m, 'factors')]
-                maps = list(flatten(maps))
-                set1, set2 = base_loop.it_space.iterset, loop.it_space.iterset
-                fused_map = [m for m in maps if set1 == m.iterset and set2 == m.toset]
-                if fused_map:
-                    fused.append((base_loop, loop, fused_map[0], common_incs[1]))
-                    return loop_chain[:loop_chain.index(loop)+1]
-                fused_map = [m for m in maps if set1 == m.toset and set2 == m.iterset]
-                if fused_map:
-                    fused.append((loop, base_loop, fused_map[0], common_incs[0]))
-                    return loop_chain[:loop_chain.index(loop)+1]
-            return []
-
-        # First, find fusible kernels
-        fusible, skip = [], []
-        for i, l in enumerate(self._loop_chain, 1):
-            if l in skip:
-                # /l/ occurs between (hard) fusible loops, let's leave it where
-                # it is for safeness
-                continue
-            skip = fuse(l, self._loop_chain[i:], fusible)
-        if not fusible:
-            return
-
-        # Then, create a suitable hard-fusion kernel
-        # The hard fused kernel will have the following structure:
-        #
-        # wrapper (args: Union(kernel1, kernel2, extra):
-        #   staging of pointers
-        #   ...
-        #   fusion (staged pointers, ..., extra)
-        #   insertion (...)
-        #
-        # Where /extra/ represents additional arguments, like the map from
-        # /kernel1/ iteration space to /kernel2/ iteration space. The /fusion/
-        # function looks like:
-        #
-        # fusion (...):
-        #   kernel1 (buffer, ...)
-        #   for i = 0 to arity:
-        #     if not already_executed[i]:
-        #       kernel2 (buffer[..], ...)
-        #
-        # Where /arity/ is the number of /kernel2/ iterations incident to
-        # /kernel1/ iterations.
-        fused = []
-        for base_loop, fuse_loop, fused_map, fused_inc_arg in fusible:
-            # Start with analyzing the kernel ASTs. Note: fusion occurs on fresh
-            # copies of the /base/ and /fuse/ ASTs. This is because the optimization
-            # of the /fused/ AST should be independent of that of individual ASTs,
-            # and subsequent cache hits for non-fused ParLoops should always retrive
-            # the original, unmodified ASTs. This is important not just for the
-            # sake of performance, but also for correctness of padding, since hard
-            # fusion changes the signature of /fuse/ (in particular, the buffers that
-            # are provided for computation on iteration spaces)
-            finder = FindInstances((ast.FunDecl, ast.PreprocessNode))
-            base, fuse = base_loop.kernel, fuse_loop.kernel
-            base_ast = dcopy(base._original_ast) if base._code else dcopy(base._ast)
-            retval = FindInstances.default_retval()
-            base_info = finder.visit(base_ast, ret=retval)
-            base_headers = base_info[ast.PreprocessNode]
-            base_fundecl = base_info[ast.FunDecl]
-            fuse_ast = dcopy(fuse._original_ast) if fuse._code else dcopy(fuse._ast)
-            retval = FindInstances.default_retval()
-            fuse_info = finder.visit(fuse_ast, ret=retval)
-            fuse_headers = fuse_info[ast.PreprocessNode]
-            fuse_fundecl = fuse_info[ast.FunDecl]
-            if len(base_fundecl) != 1 or len(fuse_fundecl) != 1:
-                raise RuntimeError("Fusing kernels, but found unexpected AST")
-            base_fundecl = base_fundecl[0]
-            fuse_fundecl = fuse_fundecl[0]
-
-            # Create /fusion/ arguments and signature
-            body = ast.Block([])
-            fusion_name = '%s_%s' % (base_fundecl.name, fuse_fundecl.name)
-            fusion_args = dcopy(base_fundecl.args + fuse_fundecl.args)
-            fusion_fundecl = ast.FunDecl(base_fundecl.ret, fusion_name, fusion_args, body)
-
-            # Filter out duplicate arguments, and append extra arguments to the fundecl
-            binding = WeakFilter().kernel_args([base_loop, fuse_loop], fusion_fundecl)
-            fusion_fundecl.args += [ast.Decl('int*', 'executed'),
-                                    ast.Decl('int*', 'fused_iters'),
-                                    ast.Decl('int', 'i')]
-
-            # Which args are actually used in /fuse/, but not in /base/ ?
-            # The gather for such arguments is moved to /fusion/, to avoid any
-            # usless LOAD from memory
-            retval = SymbolReferences.default_retval()
-            base_symbols = SymbolReferences().visit(base_fundecl.body, ret=retval)
-            retval = SymbolReferences.default_retval()
-            fuse_symbols = SymbolReferences().visit(fuse_fundecl.body, ret=retval)
-            base_funcall_syms, unshared = [], OrderedDict()
-            for arg, decl in binding.items():
-                if decl.sym.symbol in set(fuse_symbols) - set(base_symbols):
-                    base_funcall_sym = ast.Symbol('NULL')
-                    unshared.setdefault(decl, arg)
-                else:
-                    base_funcall_sym = ast.Symbol(decl.sym.symbol)
-                if arg in base_loop.args:
-                    base_funcall_syms.append(base_funcall_sym)
-            for decl, arg in unshared.items():
-                decl.typ = 'double*'
-                decl.sym.symbol = arg.c_arg_name()
-                fusion_fundecl.args.insert(fusion_fundecl.args.index(decl) + 1,
-                                           ast.Decl('int*', arg.c_map_name(0, 0)))
-
-            # Append the invocation of /base/; then, proceed with the invocation
-            # of the /fuse/ kernels
-            body.children.append(ast.FunCall(base_fundecl.name, *base_funcall_syms))
-
-            for idx in range(fused_map.arity):
-
-                fused_iter = 'fused_iters[%d]' % idx
-                fuse_funcall = ast.FunCall(fuse_fundecl.name)
-                if_cond = ast.Not(ast.Symbol('executed', (fused_iter,)))
-                if_update = ast.Assign(ast.Symbol('executed', (fused_iter,)), 1)
-                if_body = ast.Block([fuse_funcall, if_update], open_scope=True)
-                if_exec = ast.If(if_cond, [if_body])
-                body.children.extend([ast.FlatBlock('\n'), if_exec])
-
-                # Modify the /fuse/ kernel
-                # This is to take into account that many arguments are shared with
-                # /base/, so they will only staged once for /base/. This requires
-                # tweaking the way the arguments are declared and accessed in /fuse/.
-                # For example, the shared incremented array (called /buffer/ in
-                # the pseudocode in the comment above) now needs to take offsets
-                # to be sure the locations that /base/ is supposed to increment are
-                # actually accessed. The same concept apply to indirect arguments.
-                init = lambda v: '{%s}' % ', '.join([str(j) for j in v])
-                for i, fuse_loop_arg in enumerate(fuse_loop.args):
-                    fuse_kernel_arg = binding[fuse_loop_arg]
-                    buffer = '%s_vec' % fuse_kernel_arg.sym.symbol
-
-                    # How should I use the temporaries ?
-                    if fuse_loop_arg.access == INC:
-                        op = ast.Incr
-                        lvalue, rvalue = fuse_kernel_arg.sym.symbol, buffer
-                        extend_if_body = lambda body, block: body.children.extend(block)
-                        buffer_decl = ast.Decl('%s' % fuse_loop_arg.ctype, buffer)
-                    elif fuse_loop_arg.access == READ:
-                        op = ast.Assign
-                        lvalue, rvalue = buffer, fuse_kernel_arg.sym.symbol
-                        extend_if_body = lambda body, block: \
-                            [body.children.insert(0, b) for b in reversed(block)]
-                        buffer_decl = ast.Decl('%s*' % fuse_loop_arg.ctype, buffer)
-
-                    # Now handle arguments depending on their type ...
-                    if fuse_loop_arg._is_mat:
-                        # ... Handle Mats
-                        staging = []
-                        for b in fused_inc_arg._block_shape:
-                            for rc in b:
-                                lvalue = ast.Symbol(lvalue, (idx, idx),
-                                                    ((rc[0], 'j'), (rc[1], 'k')))
-                                rvalue = ast.Symbol(rvalue, ('j', 'k'))
-                                staging = ItSpace(mode=0).to_for([(0, rc[0]), (0, rc[1])],
-                                                                 ('j', 'k'),
-                                                                 [op(lvalue, rvalue)])[:1]
-                        # Set up the temporary
-                        buffer_decl.sym.rank = fuse_kernel_arg.sym.rank
-                        if fuse_loop_arg.access == INC:
-                            buffer_decl.init = ast.ArrayInit(init([init([0.0])]))
-
-                    elif fuse_loop_arg._is_indirect:
-                        # ... Handle indirect arguments. At the C level, these arguments
-                        # are of pointer type, so simple pointer arithmetic is used
-                        # to ensure the kernel accesses are to the correct locations
-                        fuse_arity = fuse_loop_arg.map.arity
-                        base_arity = fuse_arity*fused_map.arity
-                        cdim = fuse_loop_arg.data.dataset.cdim
-                        size = fuse_arity*cdim
-                        # Set the proper storage layout before invoking /fuse/
-                        ofs_vals = [[base_arity*j + k for k in range(fuse_arity)]
-                                    for j in range(cdim)]
-                        ofs_vals = [[fuse_arity*j + k for k in flatten(ofs_vals)]
-                                    for j in range(fused_map.arity)]
-                        ofs_vals = list(flatten(ofs_vals))
-                        indices = [ofs_vals[idx*size + j] for j in range(size)]
-                        # Set up the temporary and stage (gather) data into it
-                        buffer_decl.sym.rank = (size,)
-                        if fuse_loop_arg.access == INC:
-                            buffer_decl.init = ast.ArrayInit(init([0.0]))
-                            staging = [op(ast.Symbol(lvalue, (k,)), ast.Symbol(rvalue, (j,)))
-                                       for j, k in enumerate(indices)]
-                        elif fuse_kernel_arg in unshared:
-                            staging = unshared[fuse_kernel_arg].c_vec_init(False).split('\n')
-                            staging = [j for i, j in enumerate(staging) if i in indices]
-                            rvalues = [ast.FlatBlock(i.split('=')[1]) for i in staging]
-                            lvalues = [ast.Symbol(buffer, (i,)) for i in range(len(staging))]
-                            staging = [ast.Assign(i, j) for i, j in zip(lvalues, rvalues)]
-                        else:
-                            staging = [op(ast.Symbol(lvalue, (j,)), ast.Symbol(rvalue, (k,)))
-                                       for j, k in enumerate(indices)]
-
-                    else:
-                        # Nothing special to do for direct arguments
-                        continue
-
-                    # Update the If-then AST body
-                    extend_if_body(if_exec.children[0], staging)
-                    if_exec.children[0].children.insert(0, buffer_decl)
-                    fuse_funcall.children.append(ast.Symbol(buffer))
-
-            # Create a /fusion.Kernel/ object as well as the schedule
-            fused_headers = set([str(h) for h in base_headers + fuse_headers])
-            fused_ast = ast.Root([ast.PreprocessNode(h) for h in fused_headers] +
-                                 [base_fundecl, fuse_fundecl, fusion_fundecl])
-            kernels = [base, fuse]
-            loop_chain_index = (self._loop_chain.index(base_loop),
-                                self._loop_chain.index(fuse_loop))
-            # Track position of Args that need a postponed gather
-            # Can't track Args themselves as they change across different parloops
-            fargs = {fusion_args.index(i): ('postponed', False) for i in unshared.keys()}
-            fargs.update({len(set(binding.values())): ('onlymap', True)})
-            fused.append((Kernel(kernels, fused_ast, loop_chain_index), fused_map, fargs))
-
-        # Finally, generate a new schedule
-        self._schedule = HardFusionSchedule(self._name, self._schedule, fused)
-        self._loop_chain = self._schedule(self._loop_chain, only_hard=True)
-
-    def _tile(self):
-        """Tile consecutive loops over different iteration sets characterized
-        by RAW and WAR dependencies. This requires interfacing with the SLOPE
-        library."""
-
-        def inspect_set(s, insp_sets, extra_halo):
-            """Inspect the iteration set of a loop and store set info suitable
-            for SLOPE in /insp_sets/. Further, check that such iteration set has
-            a sufficiently depth halo region for correct execution in the case a
-            SLOPE MPI backend is enabled."""
-            # Get and format some iterset info
-            partitioning, superset, s_name = None, None, s.name
-            if isinstance(s, Subset):
-                superset = s.superset.name
-                s_name = "%s_ss" % s.name
-            if hasattr(s, '_partitioning'):
-                partitioning = s._partitioning
-            # If not an MPI backend, return "standard" values for core, exec, and
-            # non-exec regions (recall that SLOPE expects owned to be part of exec)
-            if slope.get_exec_mode() not in ['OMP_MPI', 'ONLY_MPI']:
-                exec_size = s.exec_size - s.core_size
-                nonexec_size = s.total_size - s.exec_size
-                infoset = s_name, s.core_size, exec_size, nonexec_size, superset
-            else:
-                if not hasattr(s, '_deep_size'):
-                    raise RuntimeError("SLOPE backend (%s) requires deep halos",
-                                       slope.get_exec_mode())
-                # Assume [1, ..., N] levels of halo depth
-                level_N = s._deep_size[-1]
-                core_size = level_N[0]
-                exec_size = level_N[2] - core_size
-                nonexec_size = level_N[3] - level_N[2]
-                if extra_halo and nonexec_size == 0:
-                    level_E = s._deep_size[-2]
-                    exec_size = level_E[2] - core_size
-                    nonexec_size = level_E[3] - level_E[2]
-                infoset = s_name, core_size, exec_size, nonexec_size, superset
-            insp_sets[infoset] = partitioning
-            return infoset
-
-        tile_size = self._options.get('tile_size', 1)
-        extra_halo = self._options.get('extra_halo', False)
-        coloring = self._options.get('coloring', 'default')
-        use_prefetch = self._options.get('use_prefetch', 0)
-        log = self._options.get('log', False)
-        rank = MPI.comm.rank
-
-        # The SLOPE inspector, which needs be populated with sets, maps,
-        # descriptors, and loop chain structure
-        inspector = slope.Inspector(self._name)
-
-        # Build inspector and argument types and values
-        # Note: we need ordered containers to be sure that SLOPE generates
-        # identical code for all ranks
-        arguments = []
-        insp_sets, insp_maps, insp_loops = OrderedDict(), OrderedDict(), []
-        for loop in self._loop_chain:
-            slope_desc = set()
-            # 1) Add sets
-            iterset = loop.it_space.iterset
-            iterset = iterset.subset if hasattr(iterset, 'subset') else iterset
-            infoset = inspect_set(iterset, insp_sets, extra_halo)
-            iterset_name, is_superset = infoset[0], infoset[4]
-            # If iterating over a subset, we fake an indirect parloop from the
-            # (iteration) subset to the superset. This allows the propagation of
-            # tiling across the hierarchy of sets (see SLOPE for further info)
-            if is_superset:
-                inspect_set(iterset.superset, insp_sets, extra_halo)
-                map_name = "%s_tosuperset" % iterset_name
-                insp_maps[iterset_name] = (map_name, iterset_name,
-                                           iterset.superset.name, iterset.indices)
-                slope_desc.add((map_name, INC._mode))
-            for a in loop.args:
-                # 2) Add access descriptors
-                maps = as_tuple(a.map, Map)
-                if not maps:
-                    # Simplest case: direct loop
-                    slope_desc.add(('DIRECT', a.access._mode))
-                else:
-                    # Add maps (there can be more than one per argument if the arg
-                    # is actually a Mat - in which case there are two maps - or if
-                    # a MixedMap) and relative descriptors
-                    for i, map in enumerate(maps):
-                        for j, m in enumerate(map):
-                            map_name = "%s%d_%d" % (m.name, i, j)
-                            insp_maps[m.name] = (map_name, m.iterset.name,
-                                                 m.toset.name, m.values_with_halo)
-                            slope_desc.add((map_name, a.access._mode))
-                            inspect_set(m.iterset, insp_sets, extra_halo)
-                            inspect_set(m.toset, insp_sets, extra_halo)
-            # 3) Add loop
-            insp_loops.append((loop.kernel.name, iterset_name, list(slope_desc)))
-        # Provide structure of loop chain to SLOPE
-        arguments.extend([inspector.add_sets(insp_sets.keys())])
-        arguments.extend([inspector.add_maps(insp_maps.values())])
-        inspector.add_loops(insp_loops)
-
-        # Set a specific tile size
-        arguments.extend([inspector.set_tile_size(tile_size)])
-
-        # Tell SLOPE the rank of the MPI process
-        arguments.extend([inspector.set_mpi_rank(rank)])
-
-        # Get type and value of additional arguments that SLOPE can exploit
-        arguments.extend(inspector.add_extra_info())
-
-        # Add any available partitioning
-        partitionings = [(s[0], v) for s, v in insp_sets.items() if v is not None]
-        arguments.extend([inspector.add_partitionings(partitionings)])
-
-        # Arguments types and values
-        argtypes, argvalues = zip(*arguments)
-
-        # Set a tile partitioning strategy
-        inspector.set_part_mode('chunk')
-
-        # Set a tile coloring strategy
-        inspector.set_coloring(coloring)
-
-        # Inform about the prefetch distance that needs be guaranteed
-        inspector.set_prefetch_halo(use_prefetch)
-
-        # Generate the C code
-        src = inspector.generate_code()
-
-        # Return type of the inspector
-        rettype = slope.Executor.meta['py_ctype_exec']
-
-        # Compiler and linker options
-        slope_dir = os.environ['SLOPE_DIR']
-        compiler = coffee.plan.compiler.get('name')
-        cppargs = slope.get_compile_opts(compiler)
-        cppargs += ['-I%s/%s' % (slope_dir, slope.get_include_dir())]
-        ldargs = ['-L%s/%s' % (slope_dir, slope.get_lib_dir()),
-                  '-l%s' % slope.get_lib_name(),
-                  '-L%s/lib' % os.environ['SLOPE_METIS'],
-                  '-lmetis', '-lrt']
-
-        # Compile and run inspector
-        fun = compilation.load(src, "cpp", "inspector", cppargs, ldargs,
-                               argtypes, rettype, compiler)
-        inspection = fun(*argvalues)
-
-        # Log the inspector output
-        if log and rank == 0:
-            filename = os.path.join("log", "%s.txt" % self._name)
-            summary = os.path.join("log", "summary.txt")
-            if not os.path.exists(os.path.dirname(filename)):
-                os.makedirs(os.path.dirname(filename))
-            with open(filename, 'w') as f, open(summary, 'a') as s:
-                # Estimate tile footprint
-                template = '| %25s | %22s | %-11s |\n'
-                f.write('*** Tile footprint ***\n')
-                f.write(template % ('iteration set', 'memory footprint (KB)', 'megaflops'))
-                f.write('-' * 68 + '\n')
-                tot_footprint, tot_flops = 0, 0
-                for loop in self._loop_chain:
-                    flops, footprint = loop.num_flops/(1000*1000), 0
-                    for arg in loop.args:
-                        dat_size = arg.data.nbytes
-                        map_size = 0 if arg._is_direct else arg.map.values_with_halo.nbytes
-                        tot_dat_size = (dat_size + map_size)/1000
-                        footprint += tot_dat_size
-                    tot_footprint += footprint
-                    f.write(template % (loop.it_space.name, str(footprint), str(flops)))
-                    tot_flops += flops
-                f.write('** Summary: %d KBytes moved, %d Megaflops performed\n' %
-                        (tot_footprint, tot_flops))
-                probSeed = 0 if MPI.parallel else len(self._loop_chain) / 2
-                probNtiles = self._loop_chain[probSeed].it_space.exec_size / tile_size or 1
-                f.write('** KB/tile: %d' % (tot_footprint/probNtiles))
-                f.write('  (Estimated: %d tiles)\n' % probNtiles)
-                f.write('-' * 68 + '\n')
-
-                # Estimate data reuse
-                template = '| %40s | %5s | %-70s |\n'
-                f.write('*** Data reuse ***\n')
-                f.write(template % ('field', 'type', 'loops'))
-                f.write('-' * 125 + '\n')
-                reuse = OrderedDict()
-                for i, loop in enumerate(self._loop_chain):
-                    for arg in loop.args:
-                        values = reuse.setdefault(arg.data, [])
-                        if i not in values:
-                            values.append(i)
-                        if arg._is_indirect:
-                            values = reuse.setdefault(arg.map, [])
-                            if i not in values:
-                                values.append(i)
-                for field, positions in reuse.items():
-                    reused_in = ', '.join('%d' % j for j in positions)
-                    field_type = 'map' if isinstance(field, Map) else 'data'
-                    f.write(template % (field.name, field_type, reused_in))
-                ideal_reuse = 0
-                for field, positions in reuse.items():
-                    size = field.values_with_halo.nbytes if isinstance(field, Map) \
-                        else field.nbytes
-                    # First position needs be cut away as it's the first touch
-                    ideal_reuse += (size/1000)*len(positions[1:])
-                out = '** Ideal reuse (i.e., no tile growth): %d / %d KBytes (%f %%)\n' % \
-                    (ideal_reuse, tot_footprint, float(ideal_reuse)*100/tot_footprint)
-                f.write(out)
-                f.write('-' * 125 + '\n')
-                s.write(out)
-
-        # Finally, get the Executor representation, to be used at executor
-        # code generation time
-        executor = slope.Executor(inspector)
-
-        kernel = Kernel(tuple(loop.kernel for loop in self._loop_chain))
-        self._schedule = TilingSchedule(self._name, self._schedule, kernel, inspection,
-                                        executor, **self._options)
-
-    @property
-    def mode(self):
-        return self._mode
-
-    @property
-    def schedule(self):
-        return self._schedule
-
-
-# Loop fusion interface
-
-class LoopChainTag(object):
-    """A special element in the trace of lazily evaluated parallel loops that
-    delimits two different Inspectors."""
-
-    def _run(self):
-        return
-
-
-def fuse(name, loop_chain, **kwargs):
-    """Apply fusion (and possibly tiling) to an iterator of :class:`ParLoop`
-    obecjts, which we refer to as ``loop_chain``. Return an iterator of
-    :class:`ParLoop` objects, in which some loops may have been fused or tiled.
-    If fusion could not be applied, return the unmodified ``loop_chain``.
-
-    .. note::
-       At the moment, the following features are not supported, in which
-       case the unmodified ``loop_chain`` is returned.
-
-        * mixed ``Datasets`` and ``Maps``;
-        * extruded ``Sets``
-
-    .. note::
-       Tiling cannot be applied if any of the following conditions verifies:
-
-        * a global reduction/write occurs in ``loop_chain``
-    """
-    # If there is nothing to fuse, just return
-    if len(loop_chain) in [0, 1]:
-        return loop_chain
-
-    # Are there _LazyMatOp objects (i.e., synch points) preventing fusion?
-    remainder = []
-    synch_points = [l for l in loop_chain if isinstance(l, base._LazyMatOp)]
-    if synch_points:
-        if len(synch_points) > 1:
-            warning("Fusing loops and found more than one synchronization point")
-        # Fuse only the sub-sequence before the first synch point
-        synch_point = loop_chain.index(synch_points[0])
-        remainder, loop_chain = loop_chain[synch_point:], loop_chain[:synch_point]
-
-    # Get an inspector for fusing this /loop_chain/. If there's a cache hit,
-    # return the fused par loops straight away. Otherwise, try to run an inspection.
-    options = {
-        'log': kwargs.get('log', False),
-        'mode': kwargs.get('mode', 'hard'),
-        'use_glb_maps': kwargs.get('use_glb_maps', False),
-        'use_prefetch': kwargs.get('use_prefetch', 0),
-        'tile_size': kwargs.get('tile_size', 1),
-        'extra_halo': kwargs.get('extra_halo', False),
-        'coloring': kwargs.get('coloring', 'default')
-    }
-    inspector = Inspector(name, loop_chain, **options)
-    if inspector._initialized:
-        return inspector.schedule(loop_chain) + remainder
-
-    # Otherwise, is the inspection legal ?
-    mode = kwargs.get('mode', 'hard')
-    force_glb = kwargs.get('force_glb', False)
-
-    # Return if there is nothing to fuse (e.g. only _LazyMatOp objects were present)
-    if len(loop_chain) in [0, 1]:
-        return loop_chain + remainder
-
-    # Skip if loops in /loop_chain/ are already /fusion/ objects: this could happen
-    # when loops had already been fused in a /loop_chain/ context
-    if any([isinstance(l, ParLoop) for l in loop_chain]):
-        return loop_chain + remainder
-
-    # Global reductions are dangerous for correctness, so avoid fusion unless the
-    # user is forcing it
-    if not force_glb and any([l._reduced_globals for l in loop_chain]):
-        return loop_chain + remainder
-
-    # Loop fusion requires modifying kernels, so ASTs must be present...
-    if not mode == 'only_tile':
-        if any([not hasattr(l.kernel, '_ast') or not l.kernel._ast for l in loop_chain]):
-            return loop_chain + remainder
-        # ...and must not be "fake" ASTs
-        if any([isinstance(l.kernel._ast, ast.FlatBlock) for l in loop_chain]):
-            return loop_chain + remainder
-
-    # Mixed still not supported
-    if any(a._is_mixed for a in flatten([l.args for l in loop_chain])):
-        return loop_chain + remainder
-
-    # Extrusion still not supported
-    if any([l.is_layered for l in loop_chain]):
-        return loop_chain + remainder
-
-    # If tiling is requested, SLOPE must be visible
-    if mode in ['tile', 'only_tile'] and not slope:
-        return loop_chain + remainder
-
-    schedule = inspector.inspect()
-    return schedule(loop_chain) + remainder
-
-
-@decorator
-def loop_chain_tag(method, self, *args, **kwargs):
-    from base import _trace
-    retval = method(self, *args, **kwargs)
-    _trace._trace.append(LoopChainTag())
-    return retval
-
-
-@contextmanager
-def sub_loop_chain():
-    from base import _trace
-    _trace._trace.append(LoopChainTag())
-
-
-@contextmanager
-def loop_chain(name, **kwargs):
-    """Analyze the sub-trace of loops lazily evaluated in this contextmanager ::
-
-        [loop_0, loop_1, ..., loop_n-1]
-
-    and produce a new sub-trace (``m <= n``) ::
-
-        [fused_loops_0, fused_loops_1, ..., fused_loops_m-1, peel_loops]
-
-    which is eventually inserted in the global trace of :class:`ParLoop` objects.
-
-    That is, sub-sequences of :class:`ParLoop` objects are potentially replaced by
-    new :class:`ParLoop` objects representing the fusion or the tiling of the
-    original trace slice.
-
-    :arg name: identifier of the loop chain
-    :arg kwargs:
-        * mode (default='tile'): the fusion/tiling mode (accepted: soft, hard,
-            tile, only_tile).
-        * tile_size: (default=1) suggest a starting average tile size.
-        * num_unroll (default=1): in a time stepping loop, the length of the loop
-            chain is given by ``num_loops * num_unroll``, where ``num_loops`` is the
-            number of loops per time loop iteration. Therefore, setting this value
-            to a number >1 enables tiling longer chains.
-        * force_glb (default=False): force tiling even in presence of global
-            reductions. In this case, the user becomes responsible of semantic
-            correctness.
-        * split_mode (default=0): split the loop chain every /split_mode/ occurrences
-            of the special object ``LoopChainTag`` in the trace, thus creating a
-            specific inspector for each slice.
-        * coloring (default='default'): set a coloring scheme for tiling. The ``default``
-            coloring should be used because it ensures correctness by construction,
-            based on the execution mode (sequential, openmp, mpi, mixed). So this
-            should be changed only if totally confident with what is going on.
-            Possible values are default, rand, omp; these are documented in detail
-            in the documentation of the SLOPE library.
-        * explicit (default=None): an iterator of 3-tuples (f, l, ts), each 3-tuple
-            indicating a sub-sequence of loops to be inspected. ``f`` and ``l``
-            represent, respectively, the first and last loop index of the sequence;
-            ``ts`` is the tile size for the sequence. This option takes precedence
-            over /split_mode/.
-        * log (default=False): output inspector and loop chain info to a file.
-        * use_glb_maps (default=False): when tiling, use the global maps provided by
-            PyOP2, rather than the ones constructed by SLOPE.
-        * use_prefetch (default=False): when tiling, try to prefetch the next iteration.
-    """
-    assert name != lazy_trace_name, "Loop chain name must differ from %s" % lazy_trace_name
-
-    num_unroll = kwargs.setdefault('num_unroll', 1)
-    tile_size = kwargs.setdefault('tile_size', 1)
-    kwargs.setdefault('use_glb_maps', False)
-    kwargs.setdefault('use_prefetch', 0)
-    kwargs.setdefault('coloring', 'default')
-    split_mode = kwargs.pop('split_mode', 0)
-    explicit = kwargs.pop('explicit', None)
-
-    # Get a snapshot of the trace before new par loops are added within this
-    # context manager
-    from base import _trace
-    stamp = list(_trace._trace)
-
-    yield
-
-    trace = _trace._trace
-    if trace == stamp:
-        return
-
-    # What's the first item /B/ that appeared in the trace /before/ entering the
-    # context manager and that still has to be executed ?
-    # The loop chain will be (B, end_of_current_trace]
-    bottom = 0
-    for i in reversed(stamp):
-        if i in trace:
-            bottom = trace.index(i) + 1
-            break
-    extracted_trace = trace[bottom:]
-
-    # Identify sub traces
-    extracted_sub_traces, sub_trace, tags = [], [], []
-    for loop in extracted_trace:
-        if not isinstance(loop, LoopChainTag):
-            sub_trace.append(loop)
-        else:
-            tags.append(loop)
-            if split_mode and len(tags) % split_mode == 0:
-                extracted_sub_traces.append(sub_trace)
-                sub_trace = []
-    if sub_trace:
-        extracted_sub_traces.append(sub_trace)
-    extracted_trace = [i for i in extracted_trace if i not in tags]
-
-    # Four possibilities: ...
-    if num_unroll < 1:
-        # 1) ... No tiling requested, but the openmp backend was set. So we still
-        # omp-ize the loops through SLOPE
-        if slope and slope.get_exec_mode() in ['OMP', 'OMP_MPI'] and tile_size > 0:
-            block_size = tile_size    # This is rather a 'block' size (no tiling)
-            options = {'mode': 'only_omp',
-                       'tile_size': block_size}
-            new_trace = [Inspector(name, [loop], **options).inspect()([loop])
-                         for loop in extracted_trace]
-            trace[bottom:] = list(flatten(new_trace))
-            _trace.evaluate_all()
-    elif explicit:
-        # 2) ... Tile over subsets of loops in the loop chain, as specified
-        # by the user through the /explicit/ list [subset1, subset2, ...]
-        prev_last = 0
-        transformed = []
-        for i, (first, last, tile_size) in enumerate(explicit):
-            sub_name = "%s_sub%d" % (name, i)
-            kwargs['tile_size'] = tile_size
-            transformed.extend(extracted_trace[prev_last:first])
-            transformed.extend(fuse(sub_name, extracted_trace[first:last+1], **kwargs))
-            prev_last = last + 1
-        transformed.extend(extracted_trace[prev_last:])
-        trace[bottom:] = transformed
-        _trace.evaluate_all()
-    elif split_mode > 0:
-        # 3) ... Tile over subsets of loops in the loop chain. The subsets have
-        # been identified by the user through /sub_loop_chain/ or /loop_chain_tag/
-        new_trace = []
-        for i, sub_loop_chain in enumerate(extracted_sub_traces):
-            sub_name = "%s_sub%d" % (name, i)
-            new_trace.append(fuse(sub_name, sub_loop_chain, **kwargs))
-        trace[bottom:] = list(flatten(new_trace))
-        _trace.evaluate_all()
-    else:
-        # 4) ... Tile over the entire loop chain, possibly unrolled as by user
-        # request of a factor = /num_unroll/
-        total_loop_chain = loop_chain.unrolled_loop_chain + extracted_trace
-        if len(total_loop_chain) / len(extracted_trace) == num_unroll:
-            bottom = trace.index(total_loop_chain[0])
-            trace[bottom:] = fuse(name, total_loop_chain, **kwargs)
-            loop_chain.unrolled_loop_chain = []
-            _trace.evaluate_all()
-        else:
-            loop_chain.unrolled_loop_chain.extend(extracted_trace)
-loop_chain.unrolled_loop_chain = []
