+++ pyop2/fusion/__init__.py
@@ -1,0 +1,1 @@
+from interface import loop_chain, loop_chain_tag  # noqa
